QEMU: Patch for Qubes stubdom
We use a lot of stubdom overrides since many of these patches and the
configuration, only make send when used as a Qubes stubdom.  An added
bonus is it avoids triggering a qemu-native rebuild.

qemu-ifup/down and most of the patches are from
QubesOS/qubes-vmm-xen-stubdom-linux

qubes-gui-protocol.h is grabbed from here:
https://github.com/QubesOS/qubes-gui-common/blob/master/include/qubes-gui-protocol.h
It is expected to be on the host filesystem, I think.

Instead of playing with PACKAGECONFIG, we just set EXTRA_OECONF as
desired.

This also depends on libseccomp, so that pulls in the layer dependency
for meta-security.

There are some warnings since the gui-agent code uses -I/usr/include.
WARNING: qemu-3.0.0-r0 do_package_qa: QA Issue: qemu: The compile log indicates that host include and/or library paths were used.
cc1: warning: include location "/usr/include" is unsafe for cross-compilation [-Wpoison-system-directories]

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>